Updated development environment variables and added caching to TradingBot
The development environment variables in .env.vault have been updated. In addition, a new caching mechanism has been introduced to the TradingBot class. This includes a walletAmountCache map for storing amounts and timestamps, as well as a walletCacheTTL for setting cache lifespan. A trendingtimeframe configuration option was also added to fetch tokens based on the timeframe set in .env file.